feat: Restrict login to staff users on staff devices
This change introduces logic to ensure that only 'Staff' users can log in on devices designated as 'StaffDevice'. Previously, all users were visible regardless of device role.

Additionally, the initial database setup now prevents the automatic creation of a default 'Admin' user if the device is configured as a 'StaffDevice', ensuring a cleaner setup for specialized roles.